requirements.txt
pandas
numpy
requests
djangorestframework